Update Collect.Position to a long
We started getting multiple errors from different integrations:
Could not convert to integer: 158963409999360. Path 'collects[0].position'
Seems like Shopify started returning a long for this field too

+++ ShopifySharp/Entities/Collect.cs
@@ -36,7 +36,7 @@ public class Collect : ShopifyObject
     /// A number specifying the manually sorted position of this product in a custom collection. The first position is 1. This value only applies when the custom collection is viewed using the Manual sort order.
     /// </summary>
     [JsonProperty("position")]
-    public int? Position { get; set; }
+    public long? Position { get; set; }
 
     /// <summary>
     /// This is the same value as position but padded with leading zeroes to make it alphanumeric-sortable.
